+++ node/internal_binding/tcp_wrap.ts
@@ -25,7 +25,7 @@
 // - https://github.com/nodejs/node/blob/master/src/tcp_wrap.h
 
 import { notImplemented } from "../_utils.ts";
-import { unreachable } from "../../_util/asserts.ts";
+import { unreachable } from "../_util/asserts.ts";
 import { ConnectionWrap } from "./connection_wrap.ts";
 import { AsyncWrap, providerType } from "./async_wrap.ts";
 import { LibuvStreamWrap } from "./stream_wrap.ts";
