+++ node/internal_binding/udp_wrap.ts
@@ -30,7 +30,7 @@ import { Buffer } from "../buffer.ts";
 import type { ErrnoException } from "../internal/errors.ts";
 import { isIP } from "../internal/net.ts";
 
-import { isLinux, isWindows } from "../../_util/os.ts";
+import { isLinux, isWindows } from "../_util/os.ts";
 
 // @ts-ignore Deno[Deno.internal] is used on purpose here
 const DenoListenDatagram = Deno[Deno.internal]?.nodeUnstable?.listenDatagram ||
