+++ node/module.ts
@@ -27,9 +27,9 @@ import nodeMods from "./module_all.ts";
 import upstreamMods from "./upstream_modules.ts";
 
 import * as path from "../path/mod.ts";
-import { assert } from "../_util/asserts.ts";
+import { assert } from "./_util/asserts.ts";
 import { fileURLToPath, pathToFileURL } from "./url.ts";
-import { isWindows } from "../_util/os.ts";
+import { isWindows } from "./_util/os.ts";
 import {
   ERR_INVALID_MODULE_SPECIFIER,
   ERR_MODULE_NOT_FOUND,
