+++ node/net.ts
@@ -87,8 +87,8 @@ import {
   PipeConnectWrap,
 } from "./internal_binding/pipe_wrap.ts";
 import { ShutdownWrap } from "./internal_binding/stream_wrap.ts";
-import { assert } from "../_util/asserts.ts";
-import { isWindows } from "../_util/os.ts";
+import { assert } from "./_util/asserts.ts";
+import { isWindows } from "./_util/os.ts";
 import { ADDRCONFIG, lookup as dnsLookup } from "./dns.ts";
 import { codeMap } from "./internal_binding/uv.ts";
 import { guessHandleType } from "./internal_binding/util.ts";
