+++ node/path/glob.ts
@@ -1,10 +1,10 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { isWindows, osType } from "../../_util/os.ts";
+import { isWindows, osType } from "../_util/os.ts";
 import { SEP, SEP_PATTERN } from "./separator.ts";
 import * as _win32 from "./win32.ts";
 import * as _posix from "./posix.ts";
-import type { OSType } from "../../_util/os.ts";
+import type { OSType } from "../_util/os.ts";
 
 const path = isWindows ? _win32 : _posix;
 const { join, normalize } = path;
